% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Alpha}
\alias{Alpha}
\title{Alpha checks for single alphabet character}
\usage{
Alpha(...)
}
\arguments{
\item{...}{additional arguments for the primitives to be parsed}
}
\description{
Alpha checks for single alphabet character
}
\examples{
Alpha()("abc")
}

