% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{AlphaNum}
\alias{AlphaNum}
\title{AlphaNum checks for a single alphanumeric character}
\usage{
AlphaNum(...)
}
\arguments{
\item{...}{additional arguments for the primitives to be parsed}
}
\description{
AlphaNum checks for a single alphanumeric character
}
\examples{
AlphaNum()("123")
AlphaNum()("abc123")
}

