% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alt}
\alias{alt}
\title{\code{alt} combinator is similar to alternation in BNF. the parser
\code{(alt(p1, p2))} recognises anything that \code{p1} or \code{p2} would.
The approach taken in this parser follows (Fairbairn86), in which either is
interpretted in a sequential (or exclusive) manner, returning the result of
the first parser to succeed, and failure if neither does.}
\usage{
alt(p1, p2)
}
\description{
\code{alt} combinator is similar to alternation in BNF. the parser
\code{(alt(p1, p2))} recognises anything that \code{p1} or \code{p2} would.
The approach taken in this parser follows (Fairbairn86), in which either is
interpretted in a sequential (or exclusive) manner, returning the result of
the first parser to succeed, and failure if neither does.
}

