% Generated by roxygen2 (4.0.2): do not edit by hand
\name{choice}
\alias{choice}
\title{choice (or else)
(+++) || Parser a -> Parser a -> Parser a
choice(returns("1"), returns("2")) ("abcdef")}
\usage{
choice(parserp, parserq)
}
\description{
choice (or else)
(+++) || Parser a -> Parser a -> Parser a
choice(returns("1"), returns("2")) ("abcdef")
}

