% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Digit}
\alias{Digit}
\title{Digit checks for single digit}
\usage{
Digit(...)
}
\arguments{
\item{...}{additional arguments for the primitives to be parsed}
}
\description{
Digit checks for single digit
}
\examples{
Digit()("123")
}

