% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%>>=\%}
\alias{\%>>=\%}
\title{Bind/then function.}
\usage{
parserp \%>>=\% parserf
}
\description{
Bind/then function.
}
\examples{
\dontrun{
( item \%>>=\% returns("123") ) ("abc")
( item() \%>>=\% item() \%>>=\% item() \%>>=\% failure() ) ("abcdefghi")
}
}

