% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%+++\%}
\alias{\%+++\%}
\title{\code{\%+++\%} is the infix notation for the \code{alt} function.}
\usage{
p1 \%+++\% p2
}
\description{
\code{\%+++\%} is the infix notation for the \code{alt} function.
}
\examples{
(item() \%+++\% succeed("2")) ("abcdef")
}

