% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%then\%}
\alias{\%then\%}
\title{\code{\%then\%} is the infix operator for the then combinator.}
\usage{
p1 \%then\% p2
}
\description{
\code{\%then\%} is the infix operator for the then combinator.
}
\examples{
(item() \%then\% succeed("123")) ("abc")
}

