% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{\%thentree\%}
\alias{\%thentree\%}
\title{\code{\%thentree\%} is the infix operator for the then combinator, and it is
the preferred way to use the \code{thentree} operator.}
\usage{
p1 \%thentree\% p2
}
\arguments{
\item{p1}{the first parser}

\item{p2}{the second parser}
}
\value{
recognises anything that \code{p1} and \code{p2} would if placed in 
  succession.
}
\description{
\code{\%thentree\%} is the infix operator for the then combinator, and it is
the preferred way to use the \code{thentree} operator.
}
\examples{
(item() \%thentree\% succeed("123")) ("abc")
}
\seealso{
\code{\link{alt}}, \code{\link{thentree}}
}

