% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{\%using\%}
\alias{\%using\%}
\title{\code{\%using\%} is the infix operator for using}
\usage{
p \%using\% f
}
\arguments{
\item{p}{is the parser to be applied}

\item{f}{is the function to be applied to each result of \code{p}.}
}
\description{
\code{\%using\%} is the infix operator for using
}
\examples{
(item() \%using\% as.numeric) ("1abc")
}
