% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{identifier}
\alias{identifier}
\title{\code{identifier} creates an identifier}
\usage{
identifier(...)
}
\arguments{
\item{...}{takes in token primitives}
}
\description{
\code{identifier} creates an identifier
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}},
  \code{\link{natural}}, \code{\link{symbol}}
}
