% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{item}
\alias{item}
\title{\code{item} is a parser that consumes the first character of the string and
returns the rest. If it cannot consume a single character from the string, it
will emit the empty list, indicating the parser has failed.}
\usage{
item(...)
}
\arguments{
\item{...}{additional arguments for the parser}
}
\description{
\code{item} is a parser that consumes the first character of the string and
returns the rest. If it cannot consume a single character from the string, it
will emit the empty list, indicating the parser has failed.
}
\examples{
item() ("abc")
item() ("")
}
