% Generated by roxygen2 (4.0.2): do not edit by hand
\name{literal}
\alias{literal}
\title{\code{literal} is a parser for single symbols. It will attempt to match the
single symbol with the first character in the string.}
\usage{
literal(char)
}
\arguments{
\item{char}{is the character to be matched}
}
\description{
\code{literal} is a parser for single symbols. It will attempt to match the
single symbol with the first character in the string.
}
\examples{
literal("a") ("abc")
}

