% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Lower}
\alias{Lower}
\title{Lower checks for single lower case character}
\usage{
Lower(...)
}
\arguments{
\item{...}{additional arguments for the primitives to be parsed}
}
\description{
Lower checks for single lower case character
}
\examples{
Lower() ("abc")
}
\seealso{
\code{\link{Digit}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}, \code{\link{symbol}}
}

