% Generated by roxygen2 (4.0.2): do not edit by hand
\name{many}
\alias{many}
\title{many matches 0 or more of pattern p
many || Parser a -> Parser [a]}
\usage{
many(p)
}
\description{
many matches 0 or more of pattern p
many || Parser a -> Parser [a]
}
\examples{
\dontrun{
many(Digit()) ("123abc")
many(Digit()) ("abc")
}
}

