% Generated by roxygen2 (4.0.2): do not edit by hand
\name{many1}
\alias{many1}
\title{many1 matches 1 or more of pattern p
many1 || Parser a -> Parser[a]}
\usage{
many1(p)
}
\description{
many1 matches 1 or more of pattern p
many1 || Parser a -> Parser[a]
}
\examples{
\dontrun{
many1(Digit()) ("123abc")
}
}

