% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{maybe}
\alias{maybe}
\title{\code{maybe} matches 0 or 1 of pattern \code{p}.  In EBNF notation, this
corresponds to a question mark ('?').}
\usage{
maybe(p)
}
\arguments{
\item{p}{is the parser to be matched 0 or 1 times.}
}
\description{
\code{maybe} matches 0 or 1 of pattern \code{p}.  In EBNF notation, this
corresponds to a question mark ('?').
}
\examples{
maybe(Digit())("123abc")
maybe(Digit())("abc123")
}
\seealso{
\code{\link{many}}, \code{\link{some}}
}

