% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{nat}
\alias{nat}
\title{\code{nat} is a parser which matches one or more numeric characters.}
\usage{
nat()
}
\description{
\code{nat} is a parser which matches one or more numeric characters.
}
\examples{
nat() ("123 + 456")
}

