% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{natural}
\alias{natural}
\title{\code{natural} creates a token parser for natural numbers}
\usage{
natural(...)
}
\arguments{
\item{...}{additional arguments for the parser}
}
\description{
\code{natural} creates a token parser for natural numbers
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{symbol}}
}

