% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Ramble}
\alias{Ramble}
\alias{Ramble-package}
\title{Ramble is a parser generator using combinatory parsers.}
\description{
Ramble allows you to write parsers in a functional manner, inspired by
Haskell's Parsec library.
}

