% Generated by roxygen2 (4.0.2): do not edit by hand
\name{returns}
\alias{returns}
\title{returns is a function which always succeeds
returns || a -> Parser a}
\usage{
returns(string)
}
\description{
returns is a function which always succeeds
returns || a -> Parser a
}
\examples{
\dontrun{
returns("1") ("abc")
}
}

