% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{satisfy}
\alias{satisfy}
\title{\code{satisfy} is a function which allows us to make parsers that recognise single symbols.}
\usage{
satisfy(p)
}
\arguments{
\item{p}{is the predicate to determine if the arbitrary symbol is a member.}
}
\description{
\code{satisfy} is a function which allows us to make parsers that recognise single symbols.
}
