% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{space}
\alias{space}
\title{\code{space} matches zero or more space characters.}
\usage{
space()
}
\description{
\code{space} matches zero or more space characters.
}
\examples{
space() ("  abc")
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}, \code{\link{symbol}}
}
