% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{SpaceCheck}
\alias{SpaceCheck}
\title{SpaceCheck checks for a single space character}
\usage{
SpaceCheck(...)
}
\arguments{
\item{...}{additional arguments for the primitives to be parsed}
}
\description{
SpaceCheck checks for a single space character
}
\examples{
SpaceCheck()(" 123")
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}},
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}, \code{\link{symbol}}
}
