% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{String}
\alias{String}
\title{\code{string} is a combinator which allows us to build parsers which
recognise strings of symbols, rather than just single symbols}
\usage{
String(string)
}
\arguments{
\item{string}{is the string to be matched}
}
\description{
\code{string} is a combinator which allows us to build parsers which
recognise strings of symbols, rather than just single symbols
}
\examples{
String("123")("123 abc")
}

