% Generated by roxygen2 (4.0.2): do not edit by hand
\name{succeed}
\alias{succeed}
\title{\code{succeed} is based on the empty string symbol in the BNF notation The
\code{succeed} parser always succeeds, without actually consuming any input
string. Since the outcome of succeed does not depend on its input, its result
value must be pre-detemined, so it is included as an extra parameter.}
\usage{
succeed(string)
}
\arguments{
\item{string}{the result value of succeed parser}
}
\description{
\code{succeed} is based on the empty string symbol in the BNF notation The
\code{succeed} parser always succeeds, without actually consuming any input
string. Since the outcome of succeed does not depend on its input, its result
value must be pre-detemined, so it is included as an extra parameter.
}
\examples{
succeed("1") ("abc")
}

