% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{symbol}
\alias{symbol}
\title{\code{symbol} creates a token for a symbol}
\usage{
symbol(xs)
}
\arguments{
\item{xs}{takes in a string to create a token}
}
\description{
\code{symbol} creates a token for a symbol
}
\examples{
symbol("[") ("  [123]")
}

