% Generated by roxygen2 (4.0.2): do not edit by hand
\name{then}
\alias{then}
\title{\code{then} combinator corresponds to sequencing in BNF. The parser
\code{(then(p1, p2))} recognises anything that \code{p1} and \code{p2} would
if placed in succession.}
\usage{
then(p1, p2)
}
\description{
\code{then} combinator corresponds to sequencing in BNF. The parser
\code{(then(p1, p2))} recognises anything that \code{p1} and \code{p2} would
if placed in succession.
}

