% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{thentree}
\alias{thentree}
\title{\code{thentree} keeps the full tree representation of the results of parsing.
Otherwise, it is identical to \code{then}.}
\usage{
thentree(p1, p2)
}
\arguments{
\item{p1}{the first parser}

\item{p2}{the second parser}
}
\value{
recognises anything that \code{p1} and \code{p2} would if placed in 
  succession.
}
\description{
\code{\%thentree\%} is the infix operator for the then combinator, and it is
the preferred way to use the \code{thentree} operator.
}
\examples{
(item() \%thentree\% succeed("123")) ("abc")
}
\seealso{
\code{\link{alt}}, \code{\link{thentree}}
}

