% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{token}
\alias{token}
\title{\code{token} is a new primitive that ignores any space before and after
applying a parser to a token.}
\usage{
token(p)
}
\arguments{
\item{p}{is the parser to have spaces stripped.}
}
\description{
\code{token} is a new primitive that ignores any space before and after
applying a parser to a token.
}
\examples{
token(ident()) ("   variable1   ")
}

