% Generated by roxygen2 (4.0.2): do not edit by hand
\name{using}
\alias{using}
\title{\code{using} combinator allows us to manipulate results from a parser, for
example building a parse tree. The parser \code{(p \%using\% f)} has the same
behaviour as the parser \code{p}, except that the function \code{f} is
applied to each of its result values.}
\usage{
using(p, f)
}
\arguments{
\item{p}{is the parser to be applied}

\item{f}{is the function to be applied to each result of \code{p}.}
}
\description{
\code{using} combinator allows us to manipulate results from a parser, for
example building a parse tree. The parser \code{(p \%using\% f)} has the same
behaviour as the parser \code{p}, except that the function \code{f} is
applied to each of its result values.
}

