% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AlphaNum}
\alias{AlphaNum}
\title{AlphaNum checks for a single alphanumeric character}
\usage{
AlphaNum(...)
}
\description{
AlphaNum checks for a single alphanumeric character
}
\examples{
\dontrun{
AlphaNum()("123")
AlphaNum()("abc123")
}
}

