% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{alt}
\alias{alt}
\title{\code{alt} combinator is similar to alternation in BNF. the parser 
\code{(alt(p1, p2))} recognises anything that \code{p1} or \code{p2} would. 
The approach taken in this parser follows (Fairbairn86), in which either is 
interpretted in a sequential (or exclusive) manner, returning the result of
the first parser to succeed, and failure if neither does.}
\usage{
alt(p1, p2)
}
\arguments{
\item{p1}{the first parser}

\item{p2}{the second parser}
}
\value{
Returns the first parser if it suceeds otherwise the second parser
}
\description{
\code{alt} combinator is similar to alternation in BNF. the parser 
\code{(alt(p1, p2))} recognises anything that \code{p1} or \code{p2} would. 
The approach taken in this parser follows (Fairbairn86), in which either is 
interpretted in a sequential (or exclusive) manner, returning the result of
the first parser to succeed, and failure if neither does.
}

