% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%alt\%}
\alias{\%alt\%}
\title{\code{\%+++\%} is the infix notation for the \code{alt} function.}
\usage{
p1 \%alt\% p2
}
\description{
\code{\%+++\%} is the infix notation for the \code{alt} function.
}
\examples{
(item() \%+++\% succeed("2")) ("abcdef")
}

