% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{\%alt\%}
\alias{\%alt\%}
\title{\code{\%alt\%} is the infix notation for the \code{alt} function.}
\usage{
p1 \%alt\% p2
}
\arguments{
\item{p1}{the first parser}

\item{p2}{the second parser}
}
\value{
Returns the first parser if it suceeds otherwise the second parser
}
\description{
\code{\%alt\%} is the infix notation for the \code{alt} function.
}
\examples{
(item() \%alt\% succeed("2")) ("abcdef")
}
