% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%+++\%}
\alias{\%+++\%}
\title{binary operand for choice
Usage:
```
(item() %+++% returns("2")) ("abcdef")
```}
\usage{
parserp \%+++\% parserq
}
\description{
binary operand for choice
Usage:
```
(item() %+++% returns("2")) ("abcdef")
```
}
\examples{
\dontrun{
(item \%+++\% returns("2")) ("abcdef")
}
}

