% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{\%then\%}
\alias{\%then\%}
\title{\code{\%then\%} is the infix operator for the then combinator.}
\usage{
p1 \%then\% p2
}
\arguments{
\item{p1}{the first parser}

\item{p2}{the second parser}
}
\value{
recognises anything that \code{p1} and \code{p2} would if placed in succession.
}
\description{
\code{\%then\%} is the infix operator for the then combinator.
}
\examples{
(item() \%then\% succeed("123")) ("abc")
}

