% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ident}
\alias{ident}
\title{\code{ident} is a parser which matches zero or more alphanumeric
characters.}
\usage{
ident()
}
\description{
\code{ident} is a parser which matches zero or more alphanumeric
characters.
}
\examples{
ident() ("variable1 = 123")
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}, \code{\link{symbol}}
}

