% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ident}
\alias{ident}
\title{\code{ident} is a parser which matches zero or more alphanumeric
characters.}
\usage{
ident()
}
\description{
\code{ident} is a parser which matches zero or more alphanumeric
characters.
}
\examples{
ident() ("variable1 = 123")
}

