% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ident}
\alias{ident}
\title{ident is identify, which is lowercase followed by zero or more alphanumeric
ident || Parser String}
\usage{
ident()
}
\description{
ident is identify, which is lowercase followed by zero or more alphanumeric
ident || Parser String
}
\examples{
\dontrun{
ident() ("variable1 = 123")
}
}

