% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{identifier}
\alias{identifier}
\title{\code{identifier} creates an identifier}
\usage{
identifier(...)
}
\arguments{
\item{...}{takes in token primitives}
}
\description{
\code{identifier} creates an identifier
}

