% Generated by roxygen2 (4.0.2): do not edit by hand
\name{item}
\alias{item}
\title{item is a parser that consumes the first character of the string and returns
the rest. If it cannot consume a single character from the string, it will
emit the empty list, indicating the parser has failed.}
\usage{
item(...)
}
\description{
item is a parser that consumes the first character of the string and returns
the rest. If it cannot consume a single character from the string, it will
emit the empty list, indicating the parser has failed.
}
\examples{
item() ("abc")
item() ("")
}

