% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{nat}
\alias{nat}
\title{\code{nat} is a parser which matches one or more numeric characters.}
\usage{
nat()
}
\description{
\code{nat} is a parser which matches one or more numeric characters.
}
\examples{
nat() ("123 + 456")
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}},
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}, \code{\link{symbol}}
}

