% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{natural}
\alias{natural}
\title{\code{natural} creates a token parser for natural numbers}
\usage{
natural(...)
}
\arguments{
\item{...}{additional arguments for the parser}
}
\description{
\code{natural} creates a token parser for natural numbers
}

