% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramble.R
\docType{package}
\name{Ramble}
\alias{Ramble}
\alias{Ramble-package}
\alias{package-ramble}
\alias{ramble}
\title{Ramble is a parser generator using combinatory parsers.}
\description{
Ramble allows you to write parsers in a functional manner, inspired by 
Haskell's Parsec library.
}

