% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Ramble}
\alias{Ramble}
\alias{Ramble-package}
\alias{package-ramble}
\alias{ramble}
\title{Ramble is a parser generator using combinatory parsers.}
\description{
Ramble allows you to write parsers in a functional manner, this is
based on Graham Hutton's paper "higher-order functions for parsing"
}

