% Generated by roxygen2 (4.0.2): do not edit by hand
\name{String}
\alias{String}
\title{String tries to match a whole string
string || String -> Parser String}
\usage{
String(x)
}
\description{
String tries to match a whole string
string || String -> Parser String
}
\examples{
\dontrun{
String("test")("test123")
}
}

