% Generated by roxygen2 (4.0.2): do not edit by hand
\name{String}
\alias{String}
\title{\code{string} is a combinator which allows us to build parsers whcih
recognise strings of symbols, rather than just single symbols}
\usage{
String(string)
}
\arguments{
\item{string}{is the string to be matched}
}
\description{
\code{string} is a combinator which allows us to build parsers whcih
recognise strings of symbols, rather than just single symbols
}
\examples{
string("123")("123 abc")
}

