% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{symbol}
\alias{symbol}
\title{\code{symbol} creates a token for a symbol}
\usage{
symbol(xs)
}
\arguments{
\item{xs}{takes in a string to create a token}
}
\description{
\code{symbol} creates a token for a symbol
}
\examples{
symbol("[") ("  [123]")
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, \code{\link{Upper}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}
}
