% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{then}
\alias{then}
\title{\code{then} combinator corresponds to sequencing in BNF. The parser 
\code{(then(p1, p2))} recognises anything that \code{p1} and \code{p2} would 
if placed in succession.}
\usage{
then(p1, p2)
}
\arguments{
\item{p1}{the first parser}

\item{p2}{the second parser}
}
\value{
recognises anything that \code{p1} and \code{p2} would if placed in 
  succession.
}
\description{
\code{\%then\%} is the infix operator for the then combinator, and it is the
preferred way to use the \code{then} operator.
}
\examples{
(item() \%then\% succeed("123")) ("abc")
}
\seealso{
\code{\link{alt}}
}

