% Generated by roxygen2 (4.0.2): do not edit by hand
\name{then}
\alias{then}
\title{then is also know as the bind function.
then || Parser a -> (a -> Parser b) -> Parser b}
\usage{
then(parserp, parserf)
}
\description{
Usage: then(p, f) would be the same as p `then` f. Alternatively,
in the piping world, it would look like:
```
(parserp(...) %>=% parserf(...)) (string)
```
}
\details{
This is also known as the bind function
\dontrun{
then(item, returns("123")) ("abc")
}
}

