% Generated by roxygen2 (4.0.2): do not edit by hand
\name{token}
\alias{token}
\title{token strips spaces as needed
token || Parser a -> Parser a}
\usage{
token(p)
}
\description{
token strips spaces as needed
token || Parser a -> Parser a
}
\examples{
\dontrun{
token(ident) ("   variable1   ")
}
}

