% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Unlist}
\alias{Unlist}
\title{Unlist is the same as unlist, but doesn't recurses all the way to
preserve the type. This function is not well optimised.}
\usage{
Unlist(a.list)
}
\arguments{
\item{a.list}{is a list to be flatten}
}
\description{
Unlist is the same as unlist, but doesn't recurses all the way to
preserve the type. This function is not well optimised.
}

