% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{Upper}
\alias{Upper}
\title{Upper checks for a single upper case character}
\usage{
Upper(...)
}
\arguments{
\item{...}{additional arguments for the primitives to be parsed}
}
\description{
Upper checks for a single upper case character
}
\examples{
Upper()("Abc")
}
\seealso{
\code{\link{Digit}}, \code{\link{Lower}}, 
  \code{\link{Alpha}}, \code{\link{AlphaNum}}, \code{\link{SpaceCheck}}, 
  \code{\link{String}}, \code{\link{ident}}, \code{\link{nat}}, 
  \code{\link{space}}, \code{\link{token}}, \code{\link{identifier}},
  \code{\link{natural}}, \code{\link{symbol}}
}
